Allow to stream/patch any property in bokehjs
